+++ src/components/Capture/Face.js
@@ -8,12 +8,17 @@ import withCrossDeviceWhenNoCamera from './withCrossDeviceWhenNoCamera'
 import GenericError from '../GenericError'
 import FallbackButton from '../Button/FallbackButton'
 import CustomFileInput from '../CustomFileInput'
-import { isDesktop, addDeviceRelatedProperties } from '~utils'
+import {
+  isDesktop,
+  addDeviceRelatedProperties,
+  getUnsupportedMobileBrowserError,
+} from '~utils'
 import { compose } from '~utils/func'
 import { randomId } from '~utils/string'
 import { validateFile } from '~utils/file'
 import { getInactiveError } from '~utils/inactiveError'
 import { localised } from '../../locales'
+import withTheme from '../Theme'
 import theme from '../Theme/style.scss'
 import style from './style.scss'
 
@@ -23,12 +28,15 @@ const defaultPayload = {
   side: null,
 }
 
+const WrappedError = withTheme(GenericError)
+
 class Face extends Component {
   static defaultProps = {
     useUploader: false,
     requestedVariant: 'standard',
     uploadFallback: true,
     useMultipleSelfieCapture: true,
+    photoCaptureFallback: true,
     snapshotInterval: 500,
   }
 
@@ -95,6 +103,7 @@ class Face extends Component {
       useMultipleSelfieCapture,
       snapshotInterval,
       uploadFallback,
+      photoCaptureFallback,
     } = this.props
     const title = translate('selfie_capture.title')
     const props = {
@@ -118,7 +127,9 @@ class Face extends Component {
     // when we finally do get its value
     if (hasCamera === null) return
 
-    if (hasCamera) {
+    const isVideoCompatible = window.MediaRecorder != null
+
+    if (hasCamera && (isVideoCompatible || photoCaptureFallback)) {
       const ariaLabelForSelfieCameraView = translate(
         'selfie_capture.frame_accessibility'
       )
@@ -131,7 +142,7 @@ class Face extends Component {
         )
       }
 
-      if (!this.props.useUploader) {
+      if (!this.props.useUploader && photoCaptureFallback) {
         return (
           <Selfie
             {...cameraProps}
@@ -142,6 +153,19 @@ class Face extends Component {
           />
         )
       }
+    }
+
+    if (
+      !isVideoCompatible &&
+      !photoCaptureFallback &&
+      requestedVariant === 'video'
+    ) {
+      return (
+        <WrappedError
+          disableNavigation={true}
+          error={{ name: getUnsupportedMobileBrowserError() }}
+        />
+      )
     }
 
     if ((this.props.useUploader || hasCamera === false) && uploadFallback) {