+++ src/demo/demoUtils.ts
@@ -52,6 +52,7 @@ export type QueryParams = {
   customisedUI?: StringifiedBoolean
   useCustomizedApiRequests?: StringifiedBoolean
   decoupleResponse?: DecoupleResponseOptions
+  photoCaptureFallback?: StringifiedBoolean
 }
 
 export type CheckData = {
@@ -183,6 +184,8 @@ export const getInitSdkOptions = (): SdkOptions => {
       uploadFallback: queryParamToValueString.uploadFallback !== 'false',
       useMultipleSelfieCapture:
         queryParamToValueString.useMultipleSelfieCapture !== 'false',
+      photoCaptureFallback:
+        queryParamToValueString.photoCaptureFallback !== 'false',
       requestedVariant:
         queryParamToValueString.faceVideo === 'true' ? 'video' : 'standard',
     },